- Adding base for ParticleSystem - Seperating 'Game' into its own thing and moving the core operations to 'Engine' - Increasing size of GameObjectID to uint64_t to accomodate much more objects (will particles really need to increase their IDs?)

+++ include/Generic/TypeAliases.h
@@ -3,6 +3,6 @@
 namespace eloo
 {
 // Unique id given to all game objects
-using GameObjectID = uint32_t;
+using GameObjectID = uint64_t;
 using DataIndex = uint16_t;
 }
